+++ com.unity.cinemachine/Samples~/Shared Assets/Scripts/SamplesHelpUI.cs
@@ -1,5 +1,5 @@
-using System;
 using UnityEngine;
+using UnityEngine.Events;
 using UnityEngine.SceneManagement;
 using UnityEngine.UIElements;
 
@@ -9,25 +9,36 @@ namespace Unity.Cinemachine.Samples
     /// Displays a button in the game view that will bring up a window with scrollable text.
     /// </summary>
     [RequireComponent(typeof(UIDocument))]
-    public class SampleHelpUI : MonoBehaviour
+    public class SamplesHelpUI : MonoBehaviour
     {
         public bool VisibleAtStart;
         public string HelpTitle;
         [TextArea(minLines: 10, maxLines: 50)]
         public string HelpText;
 
+        [Tooltip("Event sent when the help window is dismissed")]
+        public UnityEvent OnHelpDismissed = new ();
+
         VisualElement m_HelpBox, m_HelpToggleBox;
         Toggle m_HelpToggle;
 
         void OnEnable()
         {
-            // try to center vertically too (horizontally it is) alignSelf centered
             var uiDocument = GetComponent<UIDocument>();
-            
             m_HelpToggleBox = uiDocument.rootVisualElement.Q("HelpToggleBox");
+            if (m_HelpToggleBox == null)
+            {
+                Debug.LogError("Cannot find HelpToggleBox.  Is the source asset set in the UIDocument?");
+                return;
+            }
             m_HelpToggle = new Toggle("Help");
             m_HelpToggle.AddToClassList("helpToggle");
-            m_HelpToggle.RegisterValueChangedCallback(evt => m_HelpBox.visible = evt.newValue);
+            m_HelpToggle.RegisterValueChangedCallback((evt) => 
+            {
+                m_HelpBox.visible = evt.newValue;
+                if (!evt.newValue)
+                    OnHelpDismissed.Invoke();
+            });
             m_HelpToggleBox.Add(m_HelpToggle);
             
             m_HelpBox = uiDocument.rootVisualElement.Q("HelpTextBox");
@@ -45,9 +56,13 @@ namespace Unity.Cinemachine.Samples
         void OnDisable()
         {
             CloseHelpBox();
-            m_HelpToggleBox.Remove(m_HelpToggle);
+            m_HelpToggleBox?.Remove(m_HelpToggle);
         }
 
-        void CloseHelpBox() => m_HelpToggle.value = false;
+        void CloseHelpBox()
+        {
+            if (m_HelpToggle != null)
+                m_HelpToggle.value = false;
+        }
     }
 }